fix: restructure swap
