+++ monorepo/apps/example/package.json
@@ -16,7 +16,7 @@
     "@pangolindex/core": "*",
     "@pangolindex/honeycomb-provider": "*",
     "@pangolindex/shared": "*",
-    "@pangolindex/state-hooks": "*",
+    "@pangolindex/swap": "*",
     "@pangolindex/walletmodal": "*",
     "@pangolindex/wallet-connectors": "*",
     "@types/react-router-dom": "^5.3.3",
