+++ monorepo/apps/example/src/App.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import styled from 'styled-components';
 import Layout from './layout';
-import Sample from './pages/Sample';
 import Swap from './pages/Swap';
 
 const AppWrapper = styled.div`
@@ -38,11 +37,10 @@ export default function App() {
       <BodyWrapper>
         <Routes>
           <Route path="/" element={<Layout />}>
-            <Route index element={<Sample />} />
-            <Route path="sample" element={<Sample />} />
+            <Route index element={<Swap />} />
             <Route path="swap" element={<Swap />} />
 
-            <Route path="*" element={<Sample />} />
+            <Route path="*" element={<Swap />} />
           </Route>
         </Routes>
       </BodyWrapper>
