--- monorepo/apps/example/src/pages/Sample/index.tsx
@@ -1,16 +1,0 @@
-import { Button, NewsWidget } from '@pangolindex/core';
-import React from 'react';
-import { GridContainer } from './styled';
-
-function Sample() {
-  return (
-    <GridContainer>
-      <NewsWidget />
-      <Button type="button" variant="primary" loading={true}>
-        Test
-      </Button>
-    </GridContainer>
-  );
-}
-
-export default Sample;
