--- monorepo/apps/example/src/pages/Sample/styled.ts
@@ -1,13 +1,0 @@
-import { Box } from '@pangolindex/core';
-import styled from 'styled-components';
-
-export const GridContainer = styled(Box)`
-  display: grid;
-  grid-template-columns: minmax(auto, 75%) minmax(auto, 25%);
-  grid-gap: 12px;
-  padding: 10px 0px;
-  ${({ theme }) => theme.mediaWidth.upToSmall`
-    grid-template-columns: none;
-    grid-template-rows: max-content;
-  `};
-`;
