+++ monorepo/packages/swap/src/components/ConfirmSwapDrawer/index.tsx
@@ -11,7 +11,7 @@ import {
 import React, { useContext, useMemo } from 'react';
 import { AlertTriangle, ArrowDown, ArrowUpCircle } from 'react-feather';
 import { ThemeContext } from 'styled-components';
-import { Field } from 'src/state/pswap/atom';
+import { Field } from 'src/state/atom';
 import { computeTradePriceBreakdown, warningSeverity } from 'src/utils/prices';
 import SwapDetailInfo from '../SwapDetailInfo';
 import {
