+++ monorepo/packages/swap/src/components/LimitOrder/index.tsx
@@ -24,9 +24,9 @@ import { Divide, RefreshCcw, X } from 'react-feather';
 import { ThemeContext } from 'styled-components';
 import { SwapTypes } from 'src/constants';
 import { useApproveCallbackFromInputCurrencyAmount } from 'src/hooks/useApproveCallbackFromInputCurrencyAmount';
-import { LimitField, LimitNewField } from 'src/state/pswap/atom';
-import { useGelatoLimitOrdersHook } from 'src/state/pswap/hooks';
-import { useIsSelectedAEBToken, useSwapActionHandlers } from 'src/state/pswap/hooks/common';
+import { LimitField, LimitNewField } from 'src/state/atom';
+import { useGelatoLimitOrdersHook } from 'src/state/hooks';
+import { useIsSelectedAEBToken, useSwapActionHandlers } from 'src/state/hooks/common';
 import { galetoMaxAmountSpend } from 'src/utils/galetoMaxAmountSpend';
 import { wrappedGelatoCurrency } from 'src/utils/wrappedGelatoCurrency';
 import ConfirmLimitOrderDrawer from '../ConfirmLimitOrderDrawer';
