+++ monorepo/packages/swap/src/components/SearchModal/sorting.ts
