+++ monorepo/packages/swap/src/components/Settings/WarningModal.tsx
