+++ monorepo/packages/swap/src/components/TradeOption/styled.tsx
